mama-synth/mama-synth-eval/pyproject.toml
[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mama-sia-eval"
version = "0.3.0"
description = "Evaluation methods for pre- to post-contrast translation in breast DCE-MRI"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.9"
authors = [
    {name = "mama-sia-eval contributors"}
]
keywords = [
    "medical imaging",
    "breast MRI",
    "DCE-MRI",
    "image synthesis",
    "evaluation metrics",
    "deep learning",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "numpy>=1.21.0",
    "SimpleITK>=2.2.0",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
]

[project.optional-dependencies]
frd = [
    "pyradiomics>=3.0.0",
]
lpips = [
    "torch>=1.9.0",
    "lpips>=0.1.4",
]
classification = [
    "xgboost>=1.5.0",
]
training = [
    "xgboost>=1.5.0",
    "pyradiomics>=3.0.0",
    "pandas>=1.4.0",
    "openpyxl>=3.0.0",
    "tqdm>=4.60.0",
]
segmentation = [
    "nnunetv2>=2.2",
]
viz = [
    "matplotlib>=3.5.0",
    "plotly>=5.0.0",
]
web = [
    "streamlit>=1.20.0",
    "plotly>=5.0.0",
    "pandas>=1.4.0",
]
progress = [
    "tqdm>=4.60.0",
]
all = [
    "mama-sia-eval[frd,lpips,classification,training,viz,progress]",
]
dev = [
    "mama-sia-eval[all]",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "Pillow>=9.0.0",
]

[project.urls]
Repository = "https://github.com/RichardObi/mama-sia-eval"

[project.scripts]
mamasia-train = "mama_sia_eval.train_classifier:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
mypy_path = "src"
python_version = "3.9"
check_untyped_defs = true
disallow_any_generics = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
no_implicit_reexport = true

[tool.ruff]
line-length = 88
target-version = "py39"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["mama_sia_eval"]

[tool.coverage.run]
source = ["src/mama_sia_eval"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
